Added lab2.pdf for grading
Lab 2. It's when the real work begins. It's what separates the boys from the men. Not many people make it this far. Many quit and go home. But this is where we all go to prove ourselves. This is where we truly find out what we're capable of. This is where the real class begins. Once completed, it's too late to turn around and go home. You either finish the class or die trying. This is where I, once a normal college student, will become stronger than I could have ever imagined. I can only imagine how lab 3 will continue to shape my training in the art of database.