added lab4.sql
Episode IV: We Need To Go Deeper!

In this action-packed lab, we were tasked with making use of subqueries in order to further explore the majestic depths of the cap3 database. Will the gang survive long enough to discover all the secrets held within this fantastic database snapshot? Find out next week!