added lab5.sql
Episode V: The one-liner

After a hard-fought battle, I finally believe myself to have mastered, at least in part, the ability to navigate the ever-lovely cap3 database snapshot in yet another new way, this time by using JOINs. This type of query can be very useful as it can be much more concise than subqueries and can allow us to gather information from the database in different ways. Glad to finally be able to join in on the fun!