Added lab7
After our extensive use of the cap3 database snapshot in order to learn how to navigate a database, we finally got to the point where we could learn how to design our own slightly-less-beautiful database. This lesson begins with the noble art of normalization, which this lab generously allowed us to practice by getting a sub-optimal database into the gold standard of 3NF (or, even better, the platinum standard of BCNf).