fixed lab8
Lab 8 was, unfortunately, flawed in several ways, such as by not having labels on every relationship, not relating MovieActors to Actors and likewise to directors instead of just People, and not including Role in the primary key of MovieActor so that the same actor can have several roles in the same movies. Changes were made to the other files to correspond to these changes.