added lab9
Lab 9 is our third and, sadly, final venture into the world of Normalization as far as labs are concerned. In this lab, we were allowed to again design a database and create a beautiful ER diagram for information gathered about NASA workers and equipment (SPAAAAAAAACE). The attached files include said lovely ER diagram as well as a list of the functional dependencies in the database and an argument that the database is in Boyce-Codd normal form.