added lab10.sql
The (sadly) final lab in this course consists of my use of the lovely art that is the creation of stored procedures. In the case of this lab, I wrote functions that returned the prerequisites of a given course, as well as one that returned the courses for which a given course was a prerequisite.